Fix check_cuda_version and check_cudnn_version and add their tests
