+++ install/utils.py
@@ -1,0 +1,5 @@
+def print_warning(*lines):
+    print('**************************************************')
+    for line in lines:
+        print('*** WARNING: %s' % line)
+    print('**************************************************')
